move JSON files into their dedicated folder
