+++ data/example-user-consumptions.json
@@ -1,0 +1,1 @@
+[{"uuid":"c889ad58-4421-46f1-97b8-1bdeba4845ad","mealUuid":"11ec6e3f-5949-47d3-b4b3-b3d6fab5e56b","date":[2025,6,20]},{"uuid":"e44c6186-f6cb-474b-9593-9544aafdcf69","mealUuid":"f1a08046-6309-4a14-ad37-8e776d240edd","date":[2025,6,24]},{"uuid":"35d2662d-1ced-4c29-aab2-bdbcab31290a","mealUuid":"96a62ff5-1e2a-4fc1-98c8-5691423d6dfc","date":[2025,6,20]},{"uuid":"d510dcc3-0c36-417f-8781-6b63d162954a","mealUuid":"de3b8094-0e4c-44bf-98cd-801e64d606af","date":[2025,6,20]},{"uuid":"72228952-cd4d-4fae-8b4d-6d849a135039","mealUuid":"8c911f7d-6990-4898-8917-ec9bbfad6d01","date":[2025,5,28]}]
