+++ data/example-user-meals.json
@@ -1,0 +1,1 @@
+[{"uuid":"51d2bd48-fa59-43d4-a127-5be7370fdebc","title":"Tyúkhúsleves","categoryKey":"soup","typeKey":"chicken","consumptionFrequencyDays":2,"averageCostForint":3000},{"uuid":"f12a4786-c925-4913-a3f2-099e461bb84f","title":"Marhahúsleves","categoryKey":"soup","typeKey":"beef","consumptionFrequencyDays":14,"averageCostForint":5000},{"uuid":"de3b8094-0e4c-44bf-98cd-801e64d606af","title":"Zöldségleves","categoryKey":"soup","typeKey":"vegetables","consumptionFrequencyDays":5,"averageCostForint":1500},{"uuid":"d750a18f-c7fe-4f53-8fe8-169131785aff","title":"Töltöttkáposzta","categoryKey":"main-course","typeKey":"pig","consumptionFrequencyDays":30,"averageCostForint":5000},{"uuid":"c116a2d0-b132-48b2-b964-4172f748bf2e","title":"Rántott csirkemell","categoryKey":"main-course","typeKey":"chicken","consumptionFrequencyDays":7,"averageCostForint":5000},{"uuid":"f1a08046-6309-4a14-ad37-8e776d240edd","title":"Pizza","categoryKey":"main-course","typeKey":"chicken","consumptionFrequencyDays":14,"averageCostForint":4000},{"uuid":"0c928b80-74bb-468a-a649-961bf0bf8dd5","title":"Rizs","categoryKey":"side-dish","typeKey":"plants","consumptionFrequencyDays":2,"averageCostForint":1200},{"uuid":"8c911f7d-6990-4898-8917-ec9bbfad6d01","title":"Kenyér","categoryKey":"side-dish","typeKey":"plants","consumptionFrequencyDays":1,"averageCostForint":950},{"uuid":"90be57f2-2dd7-4239-8f63-c5132312e25d","title":"Pároltzöldség","categoryKey":"side-dish","typeKey":"vegetables","consumptionFrequencyDays":1,"averageCostForint":1300},{"uuid":"96a62ff5-1e2a-4fc1-98c8-5691423d6dfc","title":"Ketchup","categoryKey":"other","typeKey":"vegetables","consumptionFrequencyDays":1,"averageCostForint":800},{"uuid":"cb66739f-0d8e-4311-9d5a-520b49322e87","title":"Majonéz","categoryKey":"other","typeKey":"vegetables","consumptionFrequencyDays":1,"averageCostForint":800},{"uuid":"11ec6e3f-5949-47d3-b4b3-b3d6fab5e56b","title":"Tartármártás","categoryKey":"other","typeKey":"plants","consumptionFrequencyDays":3,"averageCostForint":1500}]
