+++ src/main/java/david/buzas/whattoeat/repositories/MealTypeRepository.java
@@ -1,7 +1,7 @@
 package david.buzas.whattoeat.repositories;
 
 import david.buzas.whattoeat.entities.MealType;
-import david.buzas.whattoeat.utils.itemmanager.JsonFileItemManager;
+import david.buzas.whattoeat.repositories.utils.JsonFileItemManager;
 
 public class MealTypeRepository extends ItemRepository<MealType> {
